requirements.txt
asgiref==3.8.1
Django==5.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
djangorestframework==3.16.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
